+++ scripts/test.sh
@@ -1,6 +1,4 @@
 reset
 ./scripts/uninstall.sh
 ./scripts/install.sh
-#cd ..
-python examples/simple/manage.py test fobi
-#cd scripts
+python examples/simple/manage.py test fobi --settings=settings_test