+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.4.35
+-------------------------------------
+2015-03-28
+
+- Fixed the issue with `allow_multiple` working incorrectly for form handler
+  plugins. Fix the `db_store` plugin as well.
+
 0.4.34
 -------------------------------------
 2015-03-27